+++ lib/common/common_options.sh
@@ -5,13 +5,14 @@ read_param "Enter name of user: " "$user_name" user_name text
 read_param "Enter shell for user: " "$user_shell" user_shell text
 read_param "Enter password: " '' passwd secret_empty
 if [[ -z $passwd ]]; then
-  var_list[passwd]="declare -gx passwd=$passwd_default"
+  var_list[passwd]="declare -gx passwd=$passwd_default" passwd="$passwd_default"
   msg_print warning "No password entered. Password set to $passwd."
 fi
 if mountpoint -q "$dir" && [[ $(findmnt -funcevo SOURCE $dir) != tmpfs ]]; then
   read_param "Do you want to generate fstab? (Y/n): " '' fstab yes_or_no
   read_param "Do you want to install bootloader? (Y/n): " '' bootloader yes_or_no
   if [[ $bootloader == "1" ]]; then
+    read_param "Enter name of bootloader (grub2): " "grub2" bootloader_name text
     if [[ -d /sys/firmware/efi/efivars ]]; then
       bootloader_type_default=uefi
     else