+++ lib/distr/archlinux/distr_actions.sh
@@ -5,6 +5,7 @@ echo 'First step is completed.'
 source ./lib/common/common_actions_1.sh
 cat ./lib/common/rootfs_scripts/pacman_setup.sh >> $dir/root/pi_s1.sh
 cat ./lib/common/rootfs_scripts/arch_setup.sh >> $dir/root/pi_s1.sh
+[[ $bootloader == "1" ]] && cat ./lib/common/rootfs_scripts/bootloader_install/$bootloader_name.sh >> $dir/root/pi_s1.sh
 
 if [[ $arch == $arch_arch ]]; then
   arch_chroot_command="chroot_rootfs auto"