+++ lib/distr/debian/distr_actions.sh
@@ -20,6 +20,7 @@ PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 source ./lib/common/common_actions_1.sh
 cat ./lib/common/rootfs_scripts/apt_setup.sh >> $dir/root/pi_s1.sh
 cat ./lib/common/rootfs_scripts/debian_setup.sh >> $dir/root/pi_s1.sh
+[[ $bootloader == "1" ]] && cat ./lib/common/rootfs_scripts/bootloader_install/$bootloader_name.sh >> $dir/root/pi_s1.sh
 chmod +x $dir/root/pi_s1.sh
 $arch_chroot_command $dir bash /root/pi_s1.sh